+++ app/api/user/admin/users/route.ts
@@ -3,7 +3,7 @@ import dbConnect from "@/app/lib/dbConnect";
 import User, { IUser } from "@/app/models/User";
 import { currentUser } from "@clerk/nextjs/server";
 import { querySchema } from "../../../../schemas/QuerySchema";
-import { errorResponse } from "@/app/utils/errorResponse";
+import { errorResponse } from "@/utils/errorResponse";
 
 export async function GET(req: Request): Promise<NextResponse> {
   await dbConnect();
